+++ tests/test_async.py
@@ -4,7 +4,9 @@ import pytest
 import pytest_asyncio.plugin  # noqa
 
 from asyncio import Future, wait_for
+from concurrent.futures import TimeoutError
 from mock import Mock
+
 from twisted.internet.defer import ensureDeferred, succeed
 
 from pyee import EventEmitter, AsyncIOEventEmitter, TwistedEventEmitter
@@ -90,6 +92,33 @@ async def test_asyncio_error(cls, event_loop):
     result = await wait_for(should_call, 0.1)
 
     assert isinstance(result, PyeeTestError)
+
+
+@pytest.mark.asyncio
+async def test_asyncio_cancellation(event_loop):
+    """Test that AsyncIOEventEmitter can handle Future cancellations"""
+
+    cancel_me = Future(loop=event_loop)
+    should_not_call = Future(loop=event_loop)
+
+    ee = AsyncIOEventEmitter(loop=event_loop)
+
+    @ee.on('event')
+    async def event_handler():
+        cancel_me.cancel()
+
+    @ee.on('error')
+    def handle_error(exc):
+        should_not_call.set_result(None)
+
+    ee.emit('event')
+
+    try:
+        await wait_for(should_not_call, 0.1)
+    except TimeoutError:
+        pass
+    else:
+        raise PyeeTestError()
 
 
 @pytest.mark.asyncio